requirements.txt
rasterio
matplotlib
numpy
torchsummary